requirements.txt
flask
flask-cors
google-generativeai
openai
anthropic
